Do not die on COPYUID response code and friends
Parsing a COPYUID response where the uid-set was interpreted as a SEQUENCE
would throw a parse error when it hit the "]".  I've addressed this
transition.

In order to avoid more surprises like this, I've also gone through all the
RFCs referenced by the IMAP Response Codes IANA registry at
http://www.iana.org/assignments/imap-response-codes/imap-response-codes.xhtml
and added additional test coverage for the interesting variants I found.

The only case that was broken is the horrible REFERRAL response code whose
payload is an IMAPURL that can look like IMAP://user;AUTH=*@SERVER2/ or
worse.  Although it's still not clear that imap-handler throwing fatal
exceptions when it encounters input it doesn't like is desirable, I felt
that totally removing all the ATOM protections was probably not the right
way to address this one.  So I special-cased REFERRAL to simply slurp the
remainder of the response code up to the ].