Don't prevent a space after an empty section being compiled (#18)
* Don't prevent a space after an empty section being compiled

In the current master, a BODY.PEEK[] followed by any other item will be concatenated together with no space between them.  needsSpace is being set to false so there is no space between the '[' and first item in the section, and needsSpace is then set back to true during walk() on that first item.  If there are no items, needsSpace is never set back to true.  This change only sets needsSpace to false if there are items in the section and the walk happens.

* Test case for empty section

Adds a test for BODY.PEEK[] with no section contents followed by FLAGS to check that a space is inserted between them